W-11: intensive E2E harness — native headless + browser (headless Chromium)
Two layers, one code path:
- In-app E2E bridge (bridge.rs, debug/test builds only): widget
  registry, state snapshot, heartbeat, panic registry, action queue;
  exposed as window.__forgecad on wasm.
- Native harness (harness.rs/ui_tests.rs): real ui_body against a bare
  egui::Context, synthetic pointer/keyboard input at real widget
  rects — 43 tests covering every button, menu, feature and the
  empty-scene edge cases. ~1.3s.
- Browser suite (e2e/, Playwright + SwiftShader WebGL2): 22 tests
  mirroring the flows plus downloads, bridge imports, screenshots.
  Debug wasm builds force the GL backend (Dawn software WebGPU loses
  the device under the multi-pass renderer).

Six real bugs the harness caught and fixed:
1. gizmo u8 color-blend overflow — panicked on every selection (the
   "click a button and the app quits" desktop crash)
2. T/R shortcut self-deadlock (read lock inside write lock)
3. params panel pushed off-screen by the tree ScrollArea
4. renderer-mutex poison cascade on viewport picks
5. STL count*50 overflow on wasm32 crashing browser imports
6. key-chord coalescing race losing Ctrl+Shift on fast input

CI: new e2e job — debug trunk build (fast, no wasm-opt), cached rust
artifacts + Playwright browsers, artifacts uploaded on failure.

+++ crates/forge-io/src/stl.rs
@@ -96,15 +96,18 @@ pub fn read_stl_bytes(data: &[u8]) -> Result<TriMesh> {
         return Err(IoError::Malformed("file too small".into()));
     }
     // Detection: a valid binary STL has exactly 84 + 50*n bytes and a
-    // plausible triangle count.
+    // plausible triangle count. 64-bit math: `count * 50` overflows a
+    // 32-bit usize on wasm when an ASCII file's header bytes parse as a
+    // bogus ~1.8e9 triangle count (the E2E import test caught this as a
+    // hard panic in the browser build).
     let count = u32::from_le_bytes(
         data[BINARY_HEADER_LEN..BINARY_HEADER_LEN + 4]
             .try_into()
             .expect("slice length 4"),
-    ) as usize;
-    let expected = BINARY_HEADER_LEN + 4 + count * 50;
-    if data.len() == expected && count > 0 {
-        read_binary(data, count)
+    ) as u64;
+    let expected = BINARY_HEADER_LEN as u64 + 4 + count * 50;
+    if data.len() as u64 == expected && count > 0 {
+        read_binary(data, count as usize)
     } else if looks_ascii(data) {
         read_ascii(data)
     } else {
